requirements.txt
amqp==5.0.6
appdirs==1.4.4
APScheduler==3.6.3
arabic-reshaper==2.1.3
argon2-cffi==20.1.0
arrow==0.15.8
asgiref==3.7.2
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.6.0
Babel==2.8.0
backcall==0.2.0
billiard==3.6.4.0
bleach==3.3.0
cached-property==1.5.1
celery==5.1.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
ConfigArgParse==1.2.3
crispy-bootstrap5==2024.2
crypto==1.4.1
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
Django==3.2.1
django-allauth==0.44.0
django-chartjs==2.2.1
django-chunked-upload==2.0.0
django-ckeditor==5.9.0
django-crispy-forms==2.1
django-environ==0.4.5
django-filter==2.3.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-password-reset==2.0
django-phonenumber-field==4.0.0
django-tinymce==3.0.2
django-twilio==0.13.1.post0
djangorestframework==3.12.4
dnspython==2.1.0
entrypoints==0.3
et-xmlfile==1.1.0
eventlet==0.32.0
filelock==3.0.12
Flask==1.1.2
Flask-MySQLdb==0.2.0
fpdf==1.7.2
future==0.18.2
gixy==0.1.20
greenlet==1.1.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
ijson==3.1.3
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kombu==5.1.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysqlclient==2.0.3
Naked==0.1.31
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
phonenumbers==8.11.3
pickleshare==0.7.5
Pillow==8.3.1
playsound==1.2.2
prometheus-client==0.9.0
prompt-toolkit==3.0.20
protobuf==3.14.0
psycopg2==2.9.1
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
Pygments==2.8.0
PyJWT==2.1.0
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
razorpay==1.2.0
reportlab==3.6.1
requests==2.26.0
requests-oauthlib==1.3.0
schedule==1.1.0
Send2Trash==1.5.0
shellescape==3.8.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
tablib==3.0.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tweepy==3.10.0
twilio==6.35.4
typing-extensions==4.10.0
tzdata==2024.1
tzlocal==2.1
urllib3==1.26.6
vine==5.0.0
virtualenv==20.4.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
